Create ACTIONS - IW CAAD and CAWT notes
This script pulls in employer name and the income withholding amount in the dialog box.  you can chose if IW type is New, Amended, or EIWO.  It creates a CAAD note with the type of IW sent, employer name, the breakdown of monthly, weekly, biweekly, and semimonthly amount on the IW notice.  A follow up CAWT for 30 days out with the same information is optional.  If there is more than one active employer you can manually change the employer name.