Rename NOTES - IW CAAD CAWT.vbs to NOTES - IW CAAD AND CAWT.vbs
This script puts a CAAD note when sending a new, updated, eiwo income withholding notice.  The amount and employer is auto populated but can be changed.  It pulls the first employer so if there is more than one employer you may have to modify.  There is an options to add a CAWT note for 30 days in the future to make sure iw starts. The iw calculation is added to caad and cawt notes.