Create ACTIONS - PIN #.vbs
This scrip has will auto fill a word document  (not attached) to send to participant with mnonline info and participant number, name, address.  There is an option to send out pin number and option to create CAAD note of the documents sent and to who.  This may not work for all counties.