Create ACTIONS - IW Calculator
This script calculates the weekly, biweekly, semimonthly and monthly amount due for a specific case. The case number will auto populate in the dialog box.  The script pulls in the monthly accrual and nonaccrual amount from CAFS into the dialog box. A pop up box will display the payment amounts that are due.  The script brings the user to the PALC screen.  The intention of the script is for users to easily calculate the IW amount with or without 20% and to easily see if the correct amount is being received.