Update ACTIONS - UNREIMBURSED UNINSURED DOCS.vbs
removed all the DIM's to clean up the code.  separated unreimbursed and uninsured documents sent to participant from unreimbursed and uninsured documents received from participant. fixed legal heading bug from using a message box to select the correct legal heading to using a dialog box. this script is for sending the documents to the participant.