Create Create unreimbursed uninsured returned docs.vbs
resolved issue #441
Separated unreimbursed uninsured documents sent to participant and documents received from participants. This script deals with process when a party sends in a request to collect unreimbursed and uninsured expenses. This script creates a caad note of documents returned, sends 2 dord documents to the non requesting party, and create jude, obbd, ncod or cpod.